Refactor Minimap component and integrate it into WorkflowEditor for improved UI navigation
